Skip an SVG test that fails depending on what other tests came before it
Running any other SVG test before this one causes it to fail. If we were to check in new
failing results, it might start passing again if some more tests are added before it that
make it the first test that is run for a particular invocation of DumpRenderTree. So the
safest thing to do for now is to skip it. <http://webkit.org/b/35700> tracks this failure.

* platform/win/Skipped: Added svg/W3C-SVG-1.1-SE/types-dom-01-b.svg.

Identifier: 67620@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@77538 268f45cc-cd09-0410-ab3c-d52691b4dbfc