Include the crashing function in the link to a crash log on Mac
Fixes <http://webkit.org/b/53780> Crash log links in results.html should include the
function that crashed on Mac

Reviewed by David Kilzer.

* Scripts/old-run-webkit-tests:
(crashLocation): Moved all the Windows code inside an if instead of using an early return,
but didn't otherwise change it. Added an if for Mac that parses the crashing function out of
the crash log.

Identifier: 67707@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@77636 268f45cc-cd09-0410-ab3c-d52691b4dbfc