<rdar://problem/8932077> REGRESSION (5.0.3-ToT): United boarding pass has wrong layout when printed
Reviewed by Adam Roben.

Windows never got patched to pass in the page height to WebCore, so when WebCore's pagination got re-written to depend on this
value, Windows got broken.  Patch setPrinting to be identical to WebKit1 on OS X and to have it take an extra minimum page height
argument.

* WebFrame.cpp:
(WebFrame::setPrinting):
(WebFrame::setInPrintingMode):
* WebFrame.h:



Identifier: 68012@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@77953 268f45cc-cd09-0410-ab3c-d52691b4dbfc