Source/WebCore: Mac OS X Services are not available for selected text in WebKit2 windows. https://bugs.webkit.org/show_bug.cgi?id=54777 <rdar://problem/8666428>
Reviewed by Adam Roben.

The changes to WebCore for this bug are limited to exposing a new
entry point in the Editor class to write to the pasteboard and
changes to the Pasteboard class to write the selection with
a given set of pasteboard types. The majority of the work
is done in WebKit2.

* WebCore.exp.in:
* editing/Editor.h:
* editing/mac/EditorMac.mm: Added entrypoint to write the
selection to the pasteboard.
(WebCore::Editor::writeSelectionToPasteboard):
* platform/Pasteboard.h:
* platform/mac/ClipboardMac.mm:
(WebCore::ClipboardMac::writeRange):
* platform/mac/PasteboardMac.mm:
(WebCore::Pasteboard::writeSelection):

Source/WebKit2: Mac OS X Services are not available for selected text in WebKit2 windows.
https://bugs.webkit.org/show_bug.cgi?id=54777
<rdar://problem/8666428>

Reviewed by Adam Roben.

This patch adds support for Mac OS X Services in WebKit2 windows.
In WKView we now call registerServicesMenuSendTypes providing the
pasteboard types supported by WebKit and we also implement the two
protocol methods required to validate the send type and write the
content to the pasteboard. Unfortunately, AppKit expects the content
to be available in the pasteboard upon return from writeSelectionToPasteboard and
this is the reason why the call to the WebProcess is synchronous.

* Shared/SelectionState.h:
(WebKit::SelectionState::SelectionState): Extended to include
isContentRichlyEditable
* Shared/mac/PasteboardTypes.h:
* Shared/mac/PasteboardTypes.mm:
(WebKit::PasteboardTypes::forSelection): Added.
* UIProcess/API/mac/WKView.mm:
(-[WKView initWithFrame:contextRef:pageGroupRef:]): Added call
to registerServicesMenuSendTypes.
(-[WKView writeSelectionToPasteboard:types:]): Added.
(-[WKView validRequestorForSendType:returnType:]): Added.
* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::writeSelectionToPasteboard):
* UIProcess/WebPageProxy.h:
* WebProcess/WebCoreSupport/WebEditorClient.cpp:
(WebKit::WebEditorClient::respondToChangedSelection): Modified to
support the new SelectionState value.
* WebProcess/WebPage/WebPage.h:
* WebProcess/WebPage/WebPage.messages.in:
* WebProcess/WebPage/mac/WebPageMac.mm:
(WebKit::WebPage::writeSelectionToPasteboard): Added synchronous
message to write the selected content to the pasteboard.



Identifier: 69414@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@79494 268f45cc-cd09-0410-ab3c-d52691b4dbfc