WKErrors need the PeerCertificateChain for certificate errors <rdar://problem/8951784> https://bugs.webkit.org/show_bug.cgi?id=56592
Reviewed by Anders Carlsson.

* Shared/WebCoreArgumentCoders.h:
Special case encoding/decoding of ResourceErrors for the mac.

* Shared/mac/PlatformCertificateInfo.h:
* Shared/mac/PlatformCertificateInfo.mm:
(WebKit::PlatformCertificateInfo::PlatformCertificateInfo):
Add constructor that take the PeerCertificateChain as CFArrayRef.

* Shared/mac/WebCoreArgumentCodersMac.mm:
(CoreIPC::encodeResourceError):
(CoreIPC::decodeResourceError):
Add encode/decode for ResourceError using the underlying NSError
and encoding/decoding as much of the userInfo as we know how to.
Right now this includes all string values and the PeerCertificateChain.



Identifier: 71261@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@81387 268f45cc-cd09-0410-ab3c-d52691b4dbfc