AX: showContextMenu not working in WK2 https://bugs.webkit.org/show_bug.cgi?id=56734
Reviewed by Darin Adler.

../WebCore:

WebKit2 is not able to directly open a context menu because the UI is in the other process.
Instead the chrome client should be used. This also means implementing showContextMenu() in WebKit1
code.

* accessibility/mac/AccessibilityObjectWrapper.mm:
(-[AccessibilityObjectWrapper accessibilityShowContextMenu]):

../WebKit/mac:

Implement showContextMenu() so accessibility related code can call this
independent of whether WK1 or WK2 is running.

* WebCoreSupport/WebChromeClient.h:
* WebCoreSupport/WebChromeClient.mm:
(WebChromeClient::showContextMenu):



Identifier: 71458@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@81619 268f45cc-cd09-0410-ab3c-d52691b4dbfc